chore(preprocessing): moved preprocessing functions folder to domain
