+++ src/app/__init__.py
@@ -1,5 +1,5 @@
 from .injections import configure_container
 
-configure_container()
+configure_container()  # noqa: RUF067
 
 __version__ = "0.1.0"
