+++ src/app/api/error_handlers/__init__.py
@@ -7,9 +7,9 @@ from app.services.exceptions import DataValidationError
 from .file_not_found import file_not_found_error_handler
 from .training import data_validation_error_handler
 
-ExceptionHandler = Callable[[Request, Exception], Response]
+ExceptionHandler = Callable[[Request, Exception], Response]  # noqa: RUF067
 
-EXCEPTION_HANDLERS: dict[type[Exception], ExceptionHandler] = {
+EXCEPTION_HANDLERS: dict[type[Exception], ExceptionHandler] = {  # noqa: RUF067
     DataValidationError: data_validation_error_handler,  # type: ignore[dict-item]
     FileNotFoundError: file_not_found_error_handler,  # type: ignore[dict-item]
 }
