+++ src/app/api/routes/__init__.py
@@ -7,7 +7,7 @@ from .prediction import PredictionRequest, PredictionResponse, prediction_router
 from .train import TrainResponse, train_router
 from .upload import UploadResponse, upload_router
 
-ROUTERS: Iterable[APIRouter] = (
+ROUTERS: Iterable[APIRouter] = (  # noqa: RUF067
     health_router,
     prediction_router,
     train_router,
